+++ internal/testing/router/router.go
@@ -1,4 +1,4 @@
-package server
+package router
 
 import (
 	"net/http"
